finally fixed the client side cookie management in HttpImportActor, which now properly recognized cookie domains and pathhs
fixed the logout route in AuthRaceRoute, which somewhat
counter-intuitively has to use the optionalCookie directive to make sure
the internal route is matched

with that, HttpImportActor can now connect to HttpServer over https

moved to Scala 2.13.2, which requires explicit toFloat/toDouble
conversions in many places

+++ race-air/src/main/scala/gov/nasa/race/air/actor/SbsImportActor.scala
@@ -90,7 +90,7 @@ class SbsDataAcquisitionThread(socket: Socket, bufLen: Int, dropDuration: Finite
                                updateFunc: TrackedObject=>Boolean,
                                dropFunc: (String,String,DateTime,Time)=>Unit) extends SocketDataAcquisitionThread(socket) {
 
-  val dropAfter = Milliseconds(dropDuration.toMillis)
+  val dropAfter = Milliseconds(dropDuration.toMillis.toInt)
   val checkAfter = dropAfter/2
   var lastDropCheck = UndefinedDateTime
